// src/api/types.generated.ts
export interface components {
    schemas: {
// ...
        FriendResponseDto: {
            /** @example 6dbb9f7c-2a6b-4c7b-8b7b-0b4e6c2e1a33 */
            id: string;
            /** @example 9f6c2fc9-936e-4dd9-9610-fd6f7da9425a */
            userId: string;
            /** @example 1a2b3c4d-936e-4dd9-9610-fd6f7da9425a */
            friendUserId: string;
            /** @example PENDING */
            status: string;
        };
// ...
        ImportContactsDto: {
            contacts: components["schemas"]["ContactDto"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}